refactor: move youtube helpers to packages/youtube shared package with tests
- Create packages/youtube/index.ts with isYouTubeUrl + getYouTubeTitleFromUrl
- Create packages/youtube/index.test.ts (10 unit tests, moved from backend)
- Create packages/youtube/vitest.config.ts and wire test script in package.json
- backend/src/services/youtube.ts now re-exports from @timehuddle/youtube
- Add @timehuddle/youtube to backend dependencies
- Import @timehuddle/youtube in AttachmentsPanel.tsx for frontend title auto-resolve
- Update root test:all script to include packages/youtube tests
- Delete redundant backend/tests/youtube.test.ts

Agent-Logs-Url: https://github.com/mieweb/timehuddle/sessions/be461b2f-c54b-4b84-b4ec-15dcdcf452fa
