+++ backend/package.json
@@ -23,7 +23,8 @@
     "dotenv": "^17.3.1",
     "fastify": "^5.8.5",
     "mongodb": "^6.14.2",
-    "nodemailer": "^8.0.5"
+    "nodemailer": "^8.0.5",
+    "@timehuddle/youtube": "*"
   },
   "devDependencies": {
     "@types/node": "^22.13.10",
