+++ package.json
@@ -1,7 +1,10 @@
 {
   "name": "timehuddle",
   "private": true,
-  "workspaces": ["packages/*", "backend"],
+  "workspaces": [
+    "packages/*",
+    "backend"
+  ],
   "engines": {
     "node": ">=24"
   },
@@ -17,7 +20,7 @@
     "format:fix": "prettier --write .",
     "typecheck": "tsc --noEmit",
     "test": "vitest run",
-    "test:all": "vitest run && npm test --prefix backend",
+    "test:all": "npm test --prefix packages/youtube && vitest run && npm test --prefix backend",
     "test:watch": "vitest",
     "screenshots": "npx playwright test e2e/screenshots.spec.ts",
     "prepare": "node scripts/prepare-husky.cjs",
